refactor(ui): extract copy_to_clipboard to ui/clipboard.rs
The 46-line cross-platform clipboard helper (arboard primary,
wl-copy fallback on Linux/FreeBSD, sandbox-aware Linux error
messaging) moves out of main.rs into ui::clipboard. Keeps the
platform cfg-gates intact.
